Fix an oversize allocation in t_bigio on 32-bit systems (#5344)
* Fix an oversize allocation in t_bigio on 32-bit systems

A test in t_bigio.c attempts to allocate more than 4 GiB of memory,
which will overflow (and wrap) the size_t type on 32-bit systems,
creating a very small allocation instead of a very large allocation.
The test then segfaults when it accesses memory outside of  the small
buffer.

The test has been fixed by limiting the buffer to 2 GiB on 32-bit
systems.