Fix protection for __builtin_expect() in H5public.h (#5840)
This feature is required to be in H5public.h, but is protected by
an HDF5 feature test macro that will soon be private. There is also
no guarantee that the compiler used to build the software has the
same support for __builtin_expect() that the application compiler
has.

This PR updates the feature test checks for __builtin_expect()

Partial fix for #5819